Backend/requirements.txt
Flask
flask-cors
openai==0.28
python-dotenv
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
SpeechRecognition
pydub
flask